migrate prop types definition to types files
