+++ components/forms/label/component.stories.tsx
@@ -1,5 +1,6 @@
 import { Story } from '@storybook/react/types-6-0';
-import Label, { LabelProps } from './component';
+import Label from './component';
+import { LabelProps } from './types';
 
 export default {
   title: 'Components/Forms/Label',
