+++ components/forms/label/types.ts
@@ -1,0 +1,6 @@
+export interface LabelProps {
+  htmlFor?: string;
+  theme?: 'dark' | 'light';
+  children: React.ReactNode;
+  className?: string;
+}
