+++ components/forms/radio/types.ts
@@ -1,0 +1,7 @@
+import { InputHTMLAttributes } from 'react';
+
+export interface RadioProps extends InputHTMLAttributes<HTMLInputElement> {
+  theme?: 'dark' | 'light';
+  input?: Record<string, unknown>;
+  meta?: Record<string, unknown>;
+}
