+++ components/forms/textarea/types.ts
@@ -1,0 +1,7 @@
+import { TextareaHTMLAttributes } from 'react';
+
+export interface TextareaProps extends TextareaHTMLAttributes<HTMLTextAreaElement> {
+  theme?: 'dark' | 'light';
+  input?: Record<string, unknown>;
+  meta?: Record<string, unknown>;
+}
